Norg internals

This page describes the internal of Norg.jl and how it parses .norg files.

Norg

Norg internals

Tokens

Scanners

Tokenize

Tokens

Norg.Tokens — Module

Provides the tokens for the tokenizer.

A Token stores its value, a TokenType and a TokenPosition

Norg.Tokens.Token — Method

 Token(line, char, type, value)

Create a Token of type type with value value at line and char number char.

Norg.Tokens.char — Method

char(x)

Return the character number in the line corresponding to position or token x.

Norg.Tokens.line — Method

line(x)

Return the line number corresponding to the position or token x.

Scanners

Norg.Scanners — Module

Provides the scanners for the tokenizer.

The role of a scanner is to recognize a sequence of characters and to produce a Token or nothing from that.

Norg.Scanners.NORG_LINE_ENDING — Constant

All the UTF-8 characters that Norg specifies as a whitespace.

Norg.Scanners.NORG_PUNCTUATION — Constant

All the UTF-8 characters that are punctuation in Norg specification. See the norg specification

Norg.Scanners.REGISTERED_TOKENTYPES — Constant

All the registered TokenType that scan will try when consuming entries.

Norg.Scanners.scan — Function

scan([[pattern,] tokentype,] input; line=nothing, charnum=nothing)

Scan the given input for the given tokentype. It will produce a Token or nothing if the input does not match the tokentype.

line and charnum are intended to give the current position in the buffer.

If no tokentype is provided, will try the ones in REGISTERED_TOKENTYPES until one returns a Token or throw an error if none succeed.

If pattern is given, then try to fit the given patter at the start of input. If pattern is :

a Char : first character must be pattern for a match.

an AbstractString : input must startswith pattern.

an AbstractArray : call scan on each element of pattern until one matches.

a Set{Char} : first character must be included in pattern.

Tokenize

Norg.Tokenize — Module

The role of this module is solely to produce a DataStructures.Queue of tokens from an input.