HTMLCodegen

Norg.Codegen.HTMLCodegen — Module

HTML code generation using HypertextLiteral.jl.

Norg.Codegen.HTMLCodegen.FootnotesLevel — Type

Controls the position where footnotes are rendered. It can be within the lowest heading level i by setting HiFootnotes or at the root of the document or directly as it appears in the Norg document.

Norg.Codegen.HTMLCodegen.HTMLLocationTarget — Type

A special target for link location, this ensure type-stability.

Norg.Codegen.HTMLCodegen.HTMLTarget — Type

HTML target to feed codegen.