Match

Norg.Match.MatchResult

Norg.Match.match_norg

Norg.Match — Module

This module exports Match.match_norg which matches token sequences to Kinds.Kind AST nodes.

Norg.Match.MatchResult — Type

Holds results of Match.match_norg. It has a Kinds.kind, that can be found, can be closing (i.e. closing an attached modifier), continued (as in "ignore this token and continue parsing"). Whether the parser should consume or not the current token is given by the consume field.

Norg.Match.match_norg — Function

match_norg([strategy], parents, tokens, i)

Find the appropriate Kinds.Kind for a token when parser is inside a parents block parsing the tokens list at index i.

Return a Match.MatchResult.