Parser

Norg.Parser.parse_norg

Norg.Parser.parse_norg

Norg.Parser.parse_norg_dispatch

Norg.Parser — Module

This module defines the Parser.parse_norg function, which builds an AST from a token list.

The role of Parser.parse_norg is to consume tokens. To do so, it relies on Match.match_norg to take decisions on how to consume tokens.

Norg.Parser.parse_norg — Function

parse_norg(strategy, tokens, i)

Try to parse the tokens sequence starting at index i using a given strategy.

Norg.Parser.parse_norg — Method

parse_norg(tokens)

Try to parse the tokens sequence as an AST.NorgDocument starting from the begining of the sequence.

Norg.Parser.parse_norg_dispatch — Method

Main dispatch utility.