Scanners

Norg.Scanners.NORG_LINE_ENDING

Norg.Scanners.NORG_PUNCTUATION

Norg.Scanners.TOKENKIND_PARSING_ORDER

Norg.Scanners.ScanResult

Norg.Scanners.scan

Norg.Scanners.scan

Norg.Scanners — Module

Provides the scanners for the tokenizer.

The role of a scanner is to recognize a sequence of characters and to produce a ScanResult.

Norg.Scanners.NORG_LINE_ENDING — Constant

All the UTF-8 characters that Norg specifies as a whitespace.

Norg.Scanners.NORG_PUNCTUATION — Constant

All the UTF-8 characters that are punctuation in Norg specification. See the norg specification

Norg.Scanners.TOKENKIND_PARSING_ORDER — Constant

All the registered Kinds.Kind that Scanners.scan will try when consuming entries.

Norg.Scanners.ScanResult — Type

Stores the result of a scanning operation. If length is 0 that means that no matching token was found.

Norg.Scanners.scan — Function

scan(pattern, input)

Scan the given input for the given pattern.

It will produce a Scanners.ScanResult.

If pattern is given, then try to fit the given patter at the start of input. If pattern is :

a ScanStrategy subtype : scan with the given strategy (e.g. Word or Whitespace)

a Kind : parse for the given kind.

an AbstractString : input must startswith pattern.

an AbstractArray : call scan on each element of pattern until one matches.

a Set{Char} : first character must be included in pattern.

Norg.Scanners.scan — Method

scan(input; line=0, charnum=0)

Scan the given input for Scanners.TOKENKIND_PARSING_ORDER until one returns a successful Scanners.ScanResult or throw an error if none succeed.

This will return a Tokens.Token.