Tokens

Norg.Tokens.Token

Norg.Tokens.Token

Norg.Tokens.TokenPosition

Norg.Tokens.char

Norg.Tokens.line

Norg.Tokenize.tokenize

Tokens

Norg.Tokens — Module

Provides the tokens for the tokenizer.

A Tokens.Token stores its value, a Kinds.Kind and a Tokens.TokenPosition

Norg.Tokens.Token — Type

A Norg Token has a Kinds.kind (i.e. K"EndLine"), a position, and a value.

See also: Tokens.TokenPosition

Norg.Tokens.Token — Method

 Token(kind, line, char, value)

Create a Token of kind kind with value value at line and char number char.

Norg.Tokens.TokenPosition — Type

Stores the position of a token in the input file (line and char).

Norg.Tokens.char — Method

char(x)

Return the character number in the line corresponding to position or token x.

Norg.Tokens.line — Method

line(x)

Return the line number corresponding to the position or token x.

Tokenize

Norg.Tokenize — Module

Produce Tokens.Token vectors from a string using tokenize.

Norg.Tokenize.tokenize — Method

tokenize(input)

Produce Tokens.Token vectors from an input string.