Norg internals

This page describes the internal of Norg.jl and how it parses .norg files.

Norg

Norg internals

Tokens

Scanners

Tokenize

AST

Match

Parser

Codegen

Tokens

Norg.Tokens — Module

Provides the tokens for the tokenizer.

A Token stores its value, a TokenType and a TokenPosition

Norg.Tokens.Token — Method

 Token(kind, line, char, value)

Create a Token of kind kind with value value at line and char number char.

Norg.Tokens.char — Method

char(x)

Return the character number in the line corresponding to position or token x.

Norg.Tokens.line — Method

line(x)

Return the line number corresponding to the position or token x.

Scanners

Norg.Scanners — Module

Provides the scanners for the tokenizer.

The role of a scanner is to recognize a sequence of characters and to produce a ScanResult.

Norg.Scanners.NORG_LINE_ENDING — Constant

All the UTF-8 characters that Norg specifies as a whitespace.

Norg.Scanners.NORG_PUNCTUATION — Constant

All the UTF-8 characters that are punctuation in Norg specification. See the norg specification

Norg.Scanners.TOKENKIND_PARSING_ORDER — Constant

All the registered TokenType that scan will try when consuming entries.

Norg.Scanners.scan — Function

scan(pattern, input)

Scan the given input for the given pattern.

It will produce a ScanResult.

If pattern is given, then try to fit the given patter at the start of input. If pattern is :

a ScanStrategy subtype : scan with the given strategy (e.g. Word or Whitespace)

a Kind : parse for the given kind.

an AbstractString : input must startswith pattern.

an AbstractArray : call scan on each element of pattern until one matches.

a Set{Char} : first character must be included in pattern.

Norg.Scanners.scan — Method

scan(input; line=0, charnum=0)

Scan the given input for TOKENKIND_PARSING_ORDER until one returns a successful ScanResult or throw an error if none succeed.

This will return a Token.

Tokenize

Norg.Tokenize — Module

Produce

AST

Norg.AST — Module

This module defines the Abstract Syntax Trees (AST) associated with the norg format.

Norg.AST.Node — Type

An AST Node has a kind (e.g. Bold), can have children Nodes, and refer to tokens in the token array.

Match

Norg.Match — Module

This module exports match_norg which matches token sequences to NodeData types.

Norg.Match.match_norg — Function

match_norg([strategy], parents, tokens, i)

Find the appropriate Kind for a token when parser is inside a parents block parsing the tokens list at index i.

Return a Norg.Match.MatchResult.

When strategy is not specified, must return a Norg.Match.MatchFound or a Norg.Match.MatchClosing.

Parser

Norg.Parser — Module

This module defines the Norg.Parser.parse_norg function, which builds an AST from a token list.

The role of Norg.Parser.parse is to consume tokens. To do so, it relies on Norg.Match.match_norg to take decisions on how to consume tokens.

Base.parse — Method

parse(AST.NorgDocument, s)

Produce an AST.NorgDocument from a string s. Calls Tokenize.tokenize.

Norg.Parser.consume_until — Method

consume_until(k, tokens, i)

Consume tokens until a token of kind k is encountered, or final token is reached.

Norg.Parser.parse_norg — Function

parse_norg(strategy, tokens, i)

Try to parse the tokens sequence starting at index i as a nodetype.

Norg.Parser.parse_norg — Method

parse_norg(tokens)

Try to parse the tokens sequence as an AST.NorgDocument starting from the begining of the sequence.

Norg.Parser.parse_norg_dispatch — Method

Main dispatch utility.

Codegen

Modules = [Norg.Codgen]